server/requirements.txt
openenv-core>=0.2.3
fastapi>=0.110.0
uvicorn>=0.29.0
openai>=1.0.0
pyyaml>=6.0
